updates
* move do_assign_call_args_ident, do_assign_call_GET_ITEM
* formatting
* move generatedFunction to constructor
