+++ src/main/java/tripleo/elijah/stages/deduce/DeduceTypes2.java
@@ -9,14 +9,11 @@
  */
 package tripleo.elijah.stages.deduce;
 
-import com.google.common.base.Predicate;
-import com.google.common.collect.Collections2;
 import io.reactivex.rxjava3.annotations.NonNull;
 import io.reactivex.rxjava3.core.Observer;
 import io.reactivex.rxjava3.disposables.Disposable;
 import io.reactivex.rxjava3.subjects.Subject;
 import org.jdeferred2.DoneCallback;
-import org.jdeferred2.Promise;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -30,29 +27,14 @@ import tripleo.elijah.lang2.SpecialVariables;
 import tripleo.elijah.stages.deduce.declarations.DeferredMember;
 import tripleo.elijah.stages.deduce.declarations.DeferredMemberFunction;
 import tripleo.elijah.stages.gen_fn.*;
-import tripleo.elijah.stages.instructions.ConstTableIA;
-import tripleo.elijah.stages.instructions.FnCallArgs;
-import tripleo.elijah.stages.instructions.IdentIA;
-import tripleo.elijah.stages.instructions.Instruction;
-import tripleo.elijah.stages.instructions.InstructionArgument;
-import tripleo.elijah.stages.instructions.InstructionName;
-import tripleo.elijah.stages.instructions.IntegerIA;
-import tripleo.elijah.stages.instructions.Label;
-import tripleo.elijah.stages.instructions.LabelIA;
-import tripleo.elijah.stages.instructions.ProcIA;
-import tripleo.elijah.stages.instructions.VariableTableType;
+import tripleo.elijah.stages.instructions.*;
 import tripleo.elijah.stages.logging.ElLog;
-import tripleo.elijah.util.Helpers;
 import tripleo.elijah.util.NotImplementedException;
 import tripleo.elijah.work.WorkJob;
 import tripleo.elijah.work.WorkList;
 import tripleo.elijah.work.WorkManager;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.regex.Pattern;
 
 /**
@@ -2616,276 +2598,6 @@ public class DeduceTypes2 {
 			return classInvocation;
 	}
 
-	private void do_assign_call_args_ident(@NotNull BaseGeneratedFunction generatedFunction,
-										   @NotNull Context ctx,
-										   @NotNull VariableTableEntry vte,
-										   int aInstructionIndex,
-										   @NotNull ProcTableEntry aPte,
-										   int aI,
-										   @NotNull TypeTableEntry aTte,
-										   @NotNull IdentExpression aExpression) {
-		final String e_text = aExpression.getText();
-		final @Nullable InstructionArgument vte_ia = generatedFunction.vte_lookup(e_text);
-//		LOG.info("10000 "+vte_ia);
-		if (vte_ia != null) {
-			final @NotNull VariableTableEntry vte1 = generatedFunction.getVarTableEntry(to_int(vte_ia));
-			final Promise<GenType, Void, Void> p = VTE_TypePromises.do_assign_call_args_ident_vte_promise(aTte, vte1);
-			@NotNull Runnable runnable = new Runnable() {
-				boolean isDone;
-
-				@Override
-				public void run() {
-					if (isDone) return;
-					final @NotNull List<TypeTableEntry> ll = getPotentialTypesVte((GeneratedFunction) generatedFunction, vte_ia);
-					doLogic(ll);
-					isDone = true;
-				}
-
-				public void doLogic(@NotNull List<TypeTableEntry> potentialTypes) {
-					assert potentialTypes.size() >= 0;
-					switch (potentialTypes.size()) {
-						case 1:
-//							tte.attached = ll.get(0).attached;
-//							vte.addPotentialType(instructionIndex, ll.get(0));
-							if (p.isResolved())
-								LOG.info(String.format("1047 (vte already resolved) %s vte1.type = %s, gf = %s, tte1 = %s %n", vte1.getName(), vte1.type, generatedFunction, potentialTypes.get(0)));
-							else {
-								final OS_Type attached = potentialTypes.get(0).getAttached();
-								if (attached == null) return;
-								switch (attached.getType()) {
-									case USER:
-										vte1.type.setAttached(attached); // !!
-										break;
-									case USER_CLASS:
-										final GenType gt = vte1.genType;
-										gt.resolved = attached;
-										vte1.resolveType(gt);
-										break;
-									default:
-										errSink.reportWarning("Unexpected value: " + attached.getType());
-//										throw new IllegalStateException("Unexpected value: " + attached.getType());
-								}
-							}
-							break;
-						case 0:
-							// README moved up here to elimiate work
-							if (p.isResolved()) {
-								System.out.printf("890-1 Already resolved type: vte1.type = %s, gf = %s %n", vte1.type, generatedFunction);
-								break;
-							}
-							LookupResultList lrl = ctx.lookup(e_text);
-							@Nullable OS_Element best = lrl.chooseBest(null);
-							if (best instanceof FormalArgListItem) {
-								@NotNull final FormalArgListItem fali = (FormalArgListItem) best;
-								final @NotNull OS_Type osType = new OS_Type(fali.typeName());
-								if (!osType.equals(vte.type.getAttached())) {
-									@NotNull TypeTableEntry tte1 = generatedFunction.newTypeTableEntry(
-											TypeTableEntry.Type.SPECIFIED, osType, fali.getNameToken(), vte1);
-									/*if (p.isResolved())
-										System.out.printf("890 Already resolved type: vte1.type = %s, gf = %s, tte1 = %s %n", vte1.type, generatedFunction, tte1);
-									else*/ {
-										final OS_Type attached = tte1.getAttached();
-										switch (attached.getType()) {
-											case USER:
-												vte1.type.setAttached(attached); // !!
-												break;
-											case USER_CLASS:
-												final GenType gt = vte1.genType;
-												gt.resolved = attached;
-												vte1.resolveType(gt);
-												break;
-											default:
-												errSink.reportWarning("2853 Unexpected value: " + attached.getType());
-//												throw new IllegalStateException("Unexpected value: " + attached.getType());
-										}
-									}
-								}
-//								vte.type = tte1;
-//								tte.attached = tte1.attached;
-//								vte.setStatus(BaseTableEntry.Status.KNOWN, best);
-							} else if (best instanceof VariableStatement) {
-								final @NotNull VariableStatement vs = (VariableStatement) best;
-								//
-								assert vs.getName().equals(e_text);
-								//
-								@Nullable InstructionArgument vte2_ia = generatedFunction.vte_lookup(vs.getName());
-								@NotNull VariableTableEntry vte2 = generatedFunction.getVarTableEntry(to_int(vte2_ia));
-								if (p.isResolved())
-									System.out.printf("915 Already resolved type: vte2.type = %s, gf = %s %n", vte1.type, generatedFunction);
-								else {
-									final GenType gt = vte1.genType;
-									final OS_Type attached = vte2.type.getAttached();
-									gt.resolved = attached;
-									vte1.resolveType(gt);
-								}
-//								vte.type = vte2.type;
-//								tte.attached = vte.type.attached;
-								vte.setStatus(BaseTableEntry.Status.KNOWN, new GenericElementHolder(best));
-								vte2.setStatus(BaseTableEntry.Status.KNOWN, new GenericElementHolder(best)); // TODO ??
-							} else {
-								int y = 2;
-								LOG.err("543 " + best.getClass().getName());
-								throw new NotImplementedException();
-							}
-							break;
-						default:
-							// TODO hopefully this works
-							final @NotNull ArrayList<TypeTableEntry> potentialTypes1 = new ArrayList<TypeTableEntry>(
-									Collections2.filter(potentialTypes, new Predicate<TypeTableEntry>() {
-										@Override
-										public boolean apply(@org.checkerframework.checker.nullness.qual.Nullable TypeTableEntry input) {
-											assert input != null;
-											return input.getAttached() != null;
-										}
-									}));
-							// prevent infinite recursion
-							if (potentialTypes1.size() < potentialTypes.size())
-								doLogic(potentialTypes1);
-							else
-								LOG.info("913 Don't know");
-							break;
-					}
-				}
-			};
-			onFinish(runnable);
-		} else {
-			int ia = generatedFunction.addIdentTableEntry(aExpression, ctx);
-			@NotNull IdentTableEntry idte = generatedFunction.getIdentTableEntry(ia);
-			idte.addPotentialType(aInstructionIndex, aTte); // TODO DotExpression??
-			final int ii = aI;
-			idte.onType(phase, new OnType() {
-				@Override
-				public void typeDeduced(@NotNull OS_Type aType) {
-					aPte.setArgType(ii, aType); // TODO does this belong here or in FunctionInvocation?
-					aTte.setAttached(aType); // since we know that tte.attached is always null here
-				}
-
-				@Override
-				public void noTypeFound() {
-					LOG.err("719 no type found "+generatedFunction.getIdentIAPathNormal(new IdentIA(ia, generatedFunction)));
-				}
-			});
-		}
-	}
-
-	private void do_assign_call_GET_ITEM(@NotNull GetItemExpression gie, TypeTableEntry tte, @NotNull BaseGeneratedFunction generatedFunction, Context ctx) {
-		try {
-			final LookupResultList lrl = DeduceLookupUtils.lookupExpression(gie.getLeft(), ctx, this);
-			final @Nullable OS_Element best = lrl.chooseBest(null);
-			if (best != null) {
-				if (best instanceof VariableStatement) { // TODO what about alias?
-					@NotNull VariableStatement vs = (VariableStatement) best;
-					String s = vs.getName();
-					@Nullable InstructionArgument vte_ia = generatedFunction.vte_lookup(s);
-					if (vte_ia != null) {
-						@NotNull VariableTableEntry vte1 = generatedFunction.getVarTableEntry(to_int(vte_ia));
-						throw new NotImplementedException();
-					} else {
-						final IdentTableEntry idte = generatedFunction.getIdentTableEntryFor(vs.getNameToken());
-						assert idte != null;
-						if (idte.type == null) {
-							final IdentIA identIA = new IdentIA(idte.getIndex(), generatedFunction);
-							resolveIdentIA_(ctx, identIA, generatedFunction, new NullFoundElement());
-						}
-						@Nullable OS_Type ty;
-						if (idte.type == null) ty = null;
-						else ty = idte.type.getAttached();
-						idte.onType(phase, new OnType() {
-							@Override public void typeDeduced(final @NotNull OS_Type ty) {
-								assert ty != null;
-								@NotNull GenType rtype = null;
-								try {
-									rtype = resolve_type(ty, ctx);
-								} catch (ResolveError resolveError) {
-	//								resolveError.printStackTrace();
-									errSink.reportError("Cant resolve " + ty); // TODO print better diagnostic
-									return;
-								}
-								if (rtype.resolved != null && rtype.resolved.getType() == OS_Type.Type.USER_CLASS) {
-									LookupResultList lrl2 = rtype.resolved.getClassOf().getContext().lookup("__getitem__");
-									@Nullable OS_Element best2 = lrl2.chooseBest(null);
-									if (best2 != null) {
-										if (best2 instanceof FunctionDef) {
-											@NotNull FunctionDef fd = (FunctionDef) best2;
-											@Nullable ProcTableEntry pte = null;
-											final IInvocation invocation = getInvocation((GeneratedFunction) generatedFunction);
-											forFunction(newFunctionInvocation(fd, pte, invocation, phase), new ForFunction() {
-												@Override
-												public void typeDecided(final @NotNull GenType aType) {
-													assert fd == generatedFunction.getFD();
-													//
-													if (idte.type == null) {
-														idte.makeType(generatedFunction, TypeTableEntry.Type.TRANSIENT, gt(aType));  // TODO expression?
-													} else
-														idte.type.setAttached(gt(aType));
-												}
-											});
-										} else {
-											throw new NotImplementedException();
-										}
-									} else {
-										throw new NotImplementedException();
-									}
-								}
-							}
-
-							@Override
-							public void noTypeFound() {
-								throw new NotImplementedException();
-							}
-						});
-						if (ty == null) {
-							@NotNull TypeTableEntry tte3 = generatedFunction.newTypeTableEntry(
-									TypeTableEntry.Type.SPECIFIED, new OS_Type(vs.typeName()), vs.getNameToken());
-							idte.type = tte3;
-//							ty = idte.type.getAttached();
-						}
-					}
-
-	//				tte.attached = new OS_FuncType((FunctionDef) best); // TODO: what is this??
-					//vte.addPotentialType(instructionIndex, tte);
-				} else if (best instanceof FormalArgListItem) {
-					final @Nullable FormalArgListItem fali = (FormalArgListItem) best;
-					String s = fali.name();
-					@Nullable InstructionArgument vte_ia = generatedFunction.vte_lookup(s);
-					if (vte_ia != null) {
-						@NotNull VariableTableEntry vte2 = generatedFunction.getVarTableEntry(to_int(vte_ia));
-
-	//					final @Nullable OS_Type ty2 = vte2.type.attached;
-						VTE_TypePromises.getItemFali(generatedFunction, ctx, vte2, this);
-	//					vte2.onType(phase, new OnType() {
-	//						@Override public void typeDeduced(final OS_Type ty2) {
-	//						}
-	//
-	//						@Override
-	//						public void noTypeFound() {
-	//							throw new NotImplementedException();
-	//						}
-	//					});
-	/*
-						if (ty2 == null) {
-							@NotNull TypeTableEntry tte3 = generatedFunction.newTypeTableEntry(
-									TypeTableEntry.Type.SPECIFIED, new OS_Type(fali.typeName()), fali.getNameToken());
-							vte2.type = tte3;
-	//						ty2 = vte2.type.attached; // TODO this is final, but why assign anyway?
-						}
-	*/
-					}
-				} else {
-					final int y=2;
-					throw new NotImplementedException();
-				}
-			} else {
-				final int y=2;
-				throw new NotImplementedException();
-			}
-		} catch (ResolveError aResolveError) {
-			aResolveError.printStackTrace();
-			int y=2;
-			throw new NotImplementedException();
-		}
-	}
-
 /*
 	private void forFunction(GeneratedFunction gf, ForFunction forFunction) {
 		phase.forFunction(this, gf, forFunction);
@@ -3758,17 +3470,6 @@ public class DeduceTypes2 {
 			return deduceTypes2.promiseExpectation(aGeneratedFunction, aName);
 		}
 
-		public void do_assign_call_args_ident(final BaseGeneratedFunction aGeneratedFunction,
-											  final Context aCtx,
-											  final VariableTableEntry aVte,
-											  final int aInstructionIndex,
-											  final ProcTableEntry aPte,
-											  final int aI,
-											  final TypeTableEntry aTte,
-											  final IdentExpression aE) {
-			deduceTypes2.do_assign_call_args_ident(aGeneratedFunction, aCtx, aVte, aInstructionIndex, aPte, aI, aTte, aE);
-		}
-
 		public OS_Element _resolveAlias2(final AliasStatement aAliasStatement) throws ResolveError {
 			return DeduceLookupUtils._resolveAlias2(aAliasStatement, deduceTypes2);
 		}
@@ -3787,10 +3488,6 @@ public class DeduceTypes2 {
 
 		public void forFunction(final FunctionInvocation aFunctionInvocation, final ForFunction aForFunction) {
 			deduceTypes2.forFunction(aFunctionInvocation, aForFunction);
-		}
-
-		public void do_assign_call_GET_ITEM(final GetItemExpression aGie, final TypeTableEntry aTte, final BaseGeneratedFunction aGeneratedFunction, final Context aCtx) {
-			deduceTypes2.do_assign_call_GET_ITEM(aGie, aTte, aGeneratedFunction, aCtx);
 		}
 
 		public void implement_calls(final BaseGeneratedFunction aGeneratedFunction, final Context aParent, final InstructionArgument aArg, final ProcTableEntry aPte, final int aInstructionIndex) {
@@ -3820,20 +3517,21 @@ public class DeduceTypes2 {
 		public ErrSink getErrSink() {
 			return deduceTypes2.errSink;
 		}
-	}
 
-	private class NullFoundElement extends FoundElement {
-		public NullFoundElement() {
-			super(DeduceTypes2.this.phase);
+		public void onFinish(final Runnable aRunnable) {
+			deduceTypes2.onFinish(aRunnable);
 		}
 
-		@Override
-		public void foundElement(final OS_Element e) {
+		public List<TypeTableEntry> getPotentialTypesVte(final GeneratedFunction aGeneratedFunction, final InstructionArgument aVte_ia) {
+			return deduceTypes2.getPotentialTypesVte(aGeneratedFunction, aVte_ia);
 		}
 
-		@Override
-		public void noFoundElement() {
+		public DeduceTypes2 get() {
+			return deduceTypes2;
+		}
 
+		public GenType resolve_type(final OS_Type aTy, final Context aCtx) throws ResolveError {
+			return deduceTypes2.resolve_type(aTy, aCtx);
 		}
 	}
 }
